+++ src/time.cc
@@ -221,7 +221,7 @@ time_t Message2Time(uint16_t r_date, uint16_t r_time)
 		<< btohs(r_time) << ")");
 
 	time_t result = ( btohs(r_date) & 0x01ff ) - 0x29;
-	result = DayToDate( result );
+	result = DayToDate( (uint16_t)result );
 	result += (time_t)( btohs(r_time)*1.77 );
 
 	dout("Message2Time result: " << ctime(&result));
